refactor after code review
