+++ src/features/notifications/ui/notification/notification.module.scss
@@ -1,4 +1,4 @@
-@use '@/app/styles/mixins' as *;
+@use 'mixins' as *;
 
 .timeAgo {
   color: var(--light-900);
