+++ src/features/profile/api/index.ts
@@ -1,4 +1,3 @@
 export * from './devices'
-export * from './notifications'
 export * from './profile'
 export * from './subscriptions'
