+++ src/features/profile/ui/index.ts
@@ -1,3 +1,2 @@
-export * from './notifications'
 export * from './profile'
 export * from './settings'
