+++ src/shared/utils/hooks/useConnectSocket.ts
@@ -1,18 +1,15 @@
 import { useEffect, useState } from 'react'
 
 import { SocketApi } from '@/app/api'
-import { Notification } from '@/features'
+import { Notification } from '@/features/notifications'
 import { WS_EVENT_PATH } from '@/shared/config'
 
 export const useConnectSocket = () => {
   const [notifications, setNotifications] = useState<Notification>()
 
-  console.log('Notifications state:', notifications)
-
   const connectSocket = () => {
     SocketApi.createConnection()
     SocketApi.socket?.on(WS_EVENT_PATH.NOTIFICATIONS, data => {
-      console.log('notifications', data)
       setNotifications(data)
     })
   }
@@ -23,7 +20,7 @@ export const useConnectSocket = () => {
     return () => {
       SocketApi.disconnect()
     }
-  }, [])
+  }, [notifications])
 
   return { notifications }
 }
